pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hoyo-assistant"
version = "1.0.0"
description = "Your personal HoYoLAB daily task automation assistant"
readme = "README.md"
authors = [
    { name = "reimuc", email = "160387885+reimuc@users.noreply.github.com" },
]
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "aiofiles",
    "aiodns",
    "aiohttp[speedups]",
    "cachetools",
    "loguru",
    "pytz",
    "PyYAML",
    "rich",
    "tenacity",
    "pydantic-settings",
    "orjson",
]

[project.optional-dependencies]
dev = [
    "mypy",
    "pre-commit",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "types-aiofiles",
    "types-cachetools",
    "types-PyYAML",
    "types-pytz",
]

[project.scripts]
hoyo-assistant = "hoyo_assistant.cli:main"

[tool.mypy]
python_version = "3.12"
strict_optional = true
no_implicit_optional = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_any_generics = true
warn_return_any = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
pretty = true
show_error_codes = true

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
# 规则集合（= flake8 + pyflakes + pycodestyle + isort + 部分 pylint）
select = ["E", "F", "W", "I", "UP", "B", "SIM", "C4"]
ignore = ["E501"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-single-line = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=hoyo_assistant --cov-report=term-missing"
testpaths = ["tests"]
asyncio_mode = "auto"
